feat(*): Implement storage
This completely implements the storage interface for Azure. There
are currently no integration tests, but I have tested this locally
with the client example from the main bindle repo. I also have not
documented anything yet as it has a bunch of steps I want to write
up in a getting started section. Also note that there is some hackery.
As the Rust SDK improves (likely from stuff we'll contribute back),
we can improve on the hackery